+++ main/src/ca/uwaterloo/flix/language/ast/PropertyError.scala
@@ -18,15 +18,16 @@ package ca.uwaterloo.flix.language.ast
 
 import ca.uwaterloo.flix.language.ast.ExecutableAst.Property
 import ca.uwaterloo.flix.language.{CompilationError, Compiler}
+import ca.uwaterloo.flix.util.Highlight._
 
 /**
   * A common super-type for verification errors.
   */
-sealed trait PropertyError extends CompilationError
+sealed trait PropertyError extends CompilationError {
+  val kind = "Verifier Error"
+}
 
 object PropertyError {
-
-  implicit val consoleCtx = Compiler.ConsoleCtx
 
   /**
     * Returns a property error for the given property `prop` under the given environment `env`.
@@ -53,202 +54,176 @@ object PropertyError {
     * An error raised to indicate that a function is not associative.
     */
   case class AssociativityError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The function is not associative.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The function is not associative.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not associative.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that a function is not commutative.
     */
   case class CommutativityError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The function is not commutative.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The function is not commutative.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not commutative.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that a partial order is not reflexive.
     */
   case class ReflexivityError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The partial order is not reflexive.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The partial order was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The partial order is not reflexive.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not reflexive.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that a partial order is not anti-symmetric.
     */
   case class AntiSymmetryError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The partial order is not anti-symmetric.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The partial order was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The partial order is not anti-symmetric.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not anti-symmetric.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that a partial order is not transitive.
     */
   case class TransitivityError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The partial order is not transitive.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The partial order was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The partial order is not transitive.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not transitive.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the least element is not smallest.
     */
   case class LeastElementError(loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The least element is not the smallest.")}
-         |
-         |The partial order was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The least element is not the smallest element.
+           |
+           |${Code(loc, "is not the smallest element.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the lub is not an upper bound.
     */
   case class UpperBoundError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The lub is not an upper bound.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The lub was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The lub is not an upper bound.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not an upper bound.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the lub is not a least upper bound.
     */
   case class LeastUpperBoundError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The lub is not a least upper bound.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The lub was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The lub is not the least upper bound.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not the least upper bound.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the greatest element is not the largest.
     */
   case class GreatestElementError(loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The greatest element is not the largest.")}
-         |
-         |The partial order was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The greatest element is not the largest element.
+           |
+           |${Code(loc, "is not the largest element.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the glb is not a lower bound.
     */
   case class LowerBoundError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The glb is not a lower bound.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The glb was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The glb is not a lower bound.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not a lower bound.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the glb is not the greatest lower bound.
     */
   case class GreatestLowerBoundError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The glb is not a greatest lower bound.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The glb was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The glb is not the greatest lower bound.
+           |
+           |Counter-example: ${Red(m.mkString(", "))}
+           |
+           |${Code(loc, "is not the greatest lower bound.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the function is not strict.
     */
   case class StrictError(loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The function is not strict.")}
-         |
-         |The function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The function is not strict.
+           |
+           |${Code(loc, "is not strict.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the function is not monotone.
     */
   case class MonotoneError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The function is not monotone.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The function is not monotone.
+           |
+           |Counter-example: ${m.mkString(", ")}
+           |
+           |${Code(loc, "is not monotone.")}
+        """.stripMargin
   }
 
 
@@ -256,32 +231,28 @@ object PropertyError {
     * An error raised to indicate that the height function may be negative.
     */
   case class HeightNonNegativeError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The height function is not non-negative.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The height function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The height function is not non-negative.
+           |
+           |Counter-example: ${m.mkString(", ")}
+           |
+           |${Code(loc, "is not non-negative.")}
+        """.stripMargin
   }
 
   /**
     * An error raised to indicate that the height function is not strictly decreasing.
     */
   case class HeightStrictlyDecreasingError(m: Map[String, String], loc: SourceLocation) extends PropertyError {
+    val source = loc.source
     val message =
-      s"""${consoleCtx.blue(s"-- VERIFIER ERROR -------------------------------------------------- ${loc.source.format}")}
-         |
-         |${consoleCtx.red(s">> The height function is not strictly decreasing.")}
-         |
-         |Counter-example: ${m.mkString(", ")}
-         |
-         |The height function was defined here:
-         |${loc.highlight}
-         """.stripMargin
+      hl"""|>> The height function is not strictly decreasing.
+           |
+           |Counter-example: ${m.mkString(", ")}
+           |
+           |${Code(loc, "is not strictly decreasing.")}
+        """.stripMargin
   }
 
 }